Remove default `Lagrange` filter
The setting made this module incompatible with GraphicsMagick at least in version 1.3.16, which is being shipped at least with Debian stable (6.0.6) and Ubuntu 12.04 (LTS) and 12.10. We prefer GraphicsMagick with the ImageMagick wrappers to ImageMagick itself because it is said to be more stable and efficient. See http://www.graphicsmagick.org/