Fixes issue with unknown data returned
For some images unknown data in the form of lines of 0 1 can be returned
which causes the algorithm for parsing identify to be thrown significantly.
We simply use the first props object instead of the popped prop object.

This is meant to act as a temp fix while we rewrite the parse for identify.